requirements.txt
pyautogui>=0.9.53
numpy>=1.21.0
opencv-python>=4.5.3
sounddevice>=0.4.3
Pillow>=8.3.0
